Add files via upload
I have used html, css and javascript in this project. where I have used html for basic structure, css for styling. And with the help of Javascript I created questions and options as soon as the participant selects the options I took the selected option and if it is correct answer I added 1 mark to the total . Atlast I projected whole score of the participant.